Reorganize uniffi_macros::export module
The previous organization no longer made sense with the recent metadata
rewrite.
