+++ uniffi_macros/src/export.rs
@@ -2,68 +2,38 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
 
-use proc_macro2::{Ident, Span, TokenStream};
-use quote::{quote, quote_spanned};
-use syn::{
-    parse::{Parse, ParseStream},
-    spanned::Spanned,
-    LitStr, Token,
-};
+use proc_macro2::TokenStream;
+use quote::quote_spanned;
+use syn::{visit_mut::VisitMut, Item, Type};
 
-pub(crate) mod metadata;
+mod attributes;
+mod item;
 mod scaffolding;
 
-pub use self::metadata::gen_metadata;
-use self::scaffolding::{gen_fn_scaffolding, gen_method_scaffolding};
-use crate::util::{either_attribute_arg, parse_comma_separated, UniffiAttribute};
+use self::{
+    attributes::ExportAttributeArguments,
+    item::ExportItem,
+    scaffolding::{gen_fn_scaffolding, gen_method_scaffolding},
+};
 
 // TODO(jplatte): Ensure no generics, …
 // TODO(jplatte): Aggregate errors instead of short-circuiting, wherever possible
 
-pub enum ExportItem {
-    Function {
-        sig: Signature,
-    },
-    Impl {
-        self_ident: Ident,
-        methods: Vec<syn::Result<Method>>,
-    },
-}
-
-pub struct Method {
-    sig: Signature,
-}
-
-pub struct Signature {
-    ident: Ident,
-    is_async: bool,
-    inputs: Vec<syn::FnArg>,
-    output: TokenStream,
-}
-
-impl Signature {
-    fn new(item: syn::Signature) -> syn::Result<Self> {
-        let output = match item.output {
-            syn::ReturnType::Default => quote! { () },
-            syn::ReturnType::Type(_, ty) => quote! { #ty },
-        };
-
-        Ok(Self {
-            ident: item.ident,
-            is_async: item.asyncness.is_some(),
-            inputs: item.inputs.into_iter().collect(),
-            output,
-        })
-    }
-}
-
-pub fn expand_export(
-    metadata: ExportItem,
-    arguments: ExportAttributeArguments,
-    mod_path: &str,
+pub(crate) fn expand_export(
+    mut item: Item,
+    args: ExportAttributeArguments,
+    mod_path: String,
 ) -> syn::Result<TokenStream> {
+    // If the input is an `impl` block, rewrite any uses of the `Self` type
+    // alias to the actual type, so we don't have to special-case it in the
+    // metadata collection or scaffolding code generation (which generates
+    // new functions outside of the `impl`).
+    rewrite_self_type(&mut item);
+
+    let metadata = ExportItem::new(item)?;
+
     match metadata {
-        ExportItem::Function { sig } => gen_fn_scaffolding(&sig, mod_path, &arguments),
+        ExportItem::Function { sig } => gen_fn_scaffolding(&sig, &mod_path, &args),
         ExportItem::Impl {
             methods,
             self_ident,
@@ -71,71 +41,61 @@ pub fn expand_export(
             let mut method_tokens = vec![];
             for method in methods {
                 let sig = method?.sig;
-                method_tokens.push(gen_method_scaffolding(
-                    &sig,
-                    mod_path,
-                    &self_ident,
-                    &arguments,
-                )?)
+                method_tokens.push(gen_method_scaffolding(&sig, &mod_path, &self_ident, &args)?)
             }
             Ok(quote_spanned! { self_ident.span() => #(#method_tokens)* })
         }
     }
 }
 
-mod kw {
-    syn::custom_keyword!(async_runtime);
-}
+/// Rewrite Self type alias usage in an impl block to the type itself.
+///
+/// For example,
+///
+/// ```ignore
+/// impl some::module::Foo {
+///     fn method(
+///         self: Arc<Self>,
+///         arg: Option<Bar<(), Self>>,
+///     ) -> Result<Self, Error> {
+///         todo!()
+///     }
+/// }
+/// ```
+///
+/// will be rewritten to
+///
+///  ```ignore
+/// impl some::module::Foo {
+///     fn method(
+///         self: Arc<some::module::Foo>,
+///         arg: Option<Bar<(), some::module::Foo>>,
+///     ) -> Result<some::module::Foo, Error> {
+///         todo!()
+///     }
+/// }
+/// ```
+pub fn rewrite_self_type(item: &mut Item) {
+    let item = match item {
+        Item::Impl(i) => i,
+        _ => return,
+    };
 
-#[derive(Default)]
-pub struct ExportAttributeArguments {
-    async_runtime: Option<AsyncRuntime>,
-}
+    struct RewriteSelfVisitor<'a>(&'a Type);
 
-impl Parse for ExportAttributeArguments {
-    fn parse(input: ParseStream<'_>) -> syn::Result<Self> {
-        parse_comma_separated(input)
-    }
-}
-
-impl UniffiAttribute for ExportAttributeArguments {
-    fn parse_one(input: ParseStream<'_>) -> syn::Result<Self> {
-        let _: kw::async_runtime = input.parse()?;
-        let _: Token![=] = input.parse()?;
-        let async_runtime = input.parse()?;
-        Ok(Self {
-            async_runtime: Some(async_runtime),
-        })
-    }
-
-    fn merge(self, other: Self) -> syn::Result<Self> {
-        Ok(Self {
-            async_runtime: either_attribute_arg(self.async_runtime, other.async_runtime)?,
-        })
-    }
-}
-
-enum AsyncRuntime {
-    Tokio(Span),
-}
-
-impl Parse for AsyncRuntime {
-    fn parse(input: ParseStream<'_>) -> syn::Result<Self> {
-        let lit: LitStr = input.parse()?;
-        match lit.value().as_str() {
-            "tokio" => Ok(Self::Tokio(lit.span())),
-            _ => Err(syn::Error::new_spanned(
-                lit,
-                "unknown async runtime, currently only `tokio` is supported",
-            )),
+    impl<'a> VisitMut for RewriteSelfVisitor<'a> {
+        fn visit_type_mut(&mut self, i: &mut Type) {
+            match i {
+                Type::Path(p) if p.qself.is_none() && p.path.is_ident("Self") => {
+                    *i = self.0.clone();
+                }
+                _ => syn::visit_mut::visit_type_mut(self, i),
+            }
         }
     }
-}
 
-impl Spanned for AsyncRuntime {
-    fn span(&self) -> Span {
-        match self {
-            AsyncRuntime::Tokio(span) => *span,
-        }
+    let mut visitor = RewriteSelfVisitor(&item.self_ty);
+    for item in &mut item.items {
+        visitor.visit_impl_item_mut(item);
     }
 }
