--- uniffi_macros/src/export/metadata/convert.rs
@@ -1,21 +1,0 @@
-/* This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at http://mozilla.org/MPL/2.0/. */
-
-use quote::ToTokens;
-
-pub(super) fn type_as_type_path(ty: &syn::Type) -> syn::Result<&syn::TypePath> {
-    match ty {
-        syn::Type::Group(g) => type_as_type_path(&g.elem),
-        syn::Type::Paren(p) => type_as_type_path(&p.elem),
-        syn::Type::Path(p) => Ok(p),
-        _ => Err(type_not_supported(ty)),
-    }
-}
-
-fn type_not_supported(ty: &impl ToTokens) -> syn::Error {
-    syn::Error::new_spanned(
-        ty,
-        "this type is not currently supported by uniffi::export in this position",
-    )
-}
